+++ xdis/bytecode.py
@@ -1,4 +1,4 @@
-#  Copyright (c) 2018 by Rocky Bernstein
+#  Copyright (c) 2018-2019 by Rocky Bernstein
 #
 #  This program is free software; you can redistribute it and/or
 #  modify it under the terms of the GNU General Public License
@@ -231,7 +231,9 @@ def get_instructions_bytes(bytecode, opc, varnames=None, names=None, constants=N
     # multiple elements on a single pass through the loop
     n = len(bytecode)
     i = 0
+    extended_arg_count  = 0
     extended_arg = 0
+    extended_arg_size = op_size(opc.EXTENDED_ARG, opc)
     while i < n:
         op = code2num(bytecode, i)
 
@@ -307,9 +309,11 @@ def get_instructions_bytes(bytecode, opc, varnames=None, names=None, constants=N
             i += 1
 
         opname = opc.opname[op]
-        inst_size = op_size(op, opc)
+        inst_size = op_size(op, opc) + (extended_arg_count * extended_arg_size)
         yield Instruction(opname, op, optype, inst_size, arg, argval, argrepr,
-                          has_arg, offset, starts_line, is_jump_target)
+                          has_arg, offset, starts_line, is_jump_target,
+                          extended_arg_count != 0)
+        extended_arg_count = extended_arg_count + 1 if op == opc.EXTENDED_ARG else 0
 
 def op_has_argument(op, opc):
     return op >= opc.HAVE_ARGUMENT
@@ -317,8 +321,6 @@ def op_has_argument(op, opc):
 def next_offset(op, opc, offset):
     return offset + instruction_size(op, opc)
 
-# FIXME: this would better be called an instr_size
-# since it is about instructions, not opcodes
 def instruction_size(op, opc):
     """For a given opcode, `op`, in opcode module `opc`,
     return the size, in bytes, of an `op` instruction.
@@ -341,7 +343,7 @@ def instruction_size(op, opc):
 op_size = instruction_size
 
 _Instruction = namedtuple("_Instruction",
-     "opname opcode optype inst_size arg argval argrepr has_arg offset starts_line is_jump_target")
+     "opname opcode optype inst_size arg argval argrepr has_arg offset starts_line is_jump_target has_extended_arg")
 
 def from_traceback(cls, tb):
     """ Construct a Bytecode from the given traceback """
@@ -368,6 +370,8 @@ class Instruction(_Instruction):
          offset - start index of operation within bytecode sequence
          starts_line - line started by this opcode (if any), otherwise None
          is_jump_target - True if other code jumps to here, otherwise False
+         has_extended_arg - True if the instruction was built from EXTENDED_ARG
+                            opcodes
     """
     # FIXME: remove has_arg from initialization but keep it as a field.
 
@@ -467,7 +471,7 @@ class Instruction(_Instruction):
     # FIXME: figure out how to do disassemble passing in opnames
 
 class Bytecode(object):
-    """The bytecode operations of a piece of code
+    """Bytecode operations involving a Python code object.
 
     Instantiate this with a function, method, string of code, or a code object
     (as returned by compile()).